requirements.txt
certifi==2017.7.27.1
chardet==3.0.4
Django==1.11.4
django-bootstrap3==9.0.0
django-contrib-comments==1.8.0
django-crispy-forms==1.6.1
django-fluent-comments==1.4.3
django-haystack==2.6.1
django-ipware==1.1.6
django-tag-parser==3.1
django-threadedcomments==1.1
idna==2.6
jieba==0.38
Markdown==2.6.9
Pygments==2.2.0
python-akismet==0.3.0
pytz==2017.2
requests==2.18.4
urllib3==1.22
Whoosh==2.7.4
